update geonames scenario to fetch URI using https instead of http
geonames switched to use https with their URIs instead of http.  This means that the extraction process that depended on the subject URI matching the requested URI failed to find any metadata matching the ldpaths.

Consideration:

Should QA be able to find data in the graph using either/both http and https versions of the requested URI?