Remove failing tests as they are due to changes at OCLC
OCLC moved most events to be meetings.  As part of this move, the linked data search results only return `rdf:type Meeting` even if the term has `rdf:type Event Name`.

For now, tests are being disabled.

Issue # records the need to revisit this as the linked data at OCLC continues to evolve.